+++ src/ui/live/ingameui/HomePage.cpp
@@ -168,7 +168,7 @@ void HomePage::Render(float elapsedMs)
       ImGui::PushStyleVar(ImGuiStyleVar_Alpha, 0.666f * (1.f - fabs(GetOpenCloseAnimPos())));
       ImGui::SetNextWindowPos(ImVec2(0, 0));
       ImGui::SetNextWindowSize(ImGui::GetIO().DisplaySize);
-      ImGui::Begin((std::to_string(reinterpret_cast<uint64_t>(this)) + ".back").c_str(), nullptr, window_flags);
+      ImGui::Begin((std::to_string(reinterpret_cast<size_t>(this)) + ".back").c_str(), nullptr, window_flags);
       m_player->m_liveUI->SetMarkdownStartId(ImGui::GetItemID());
       ImGui::Markdown(info.str().c_str(), info.str().length(), m_player->m_liveUI->GetMarkdownConfig());
       ImGui::End();
